Replace server with simplified WebSocket implementation
- Clean 3-layer architecture: Kernel → Session → WebSocket
- Switched from xeus-python to IPython kernel (python3)
- Natural interrupt handling via kernel's KeyboardInterrupt error
- Removed test files and investigation artifacts

Architecture:
- Kernel: Low-level process management (start/stop/restart/interrupt/execute)
- Session: High-level execution coordination (parsing, task management, message processing)
- WebSocket: Thin message router

Key improvements:
- No special interrupt-ack or execution-cancelled messages
- Kernel naturally sends KeyboardInterrupt error on interrupt
- Simpler, cleaner codebase
- Better separation of concerns
